% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_functions_files.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Create a map of the count of accidents by giving information of year and state number}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{an character or integer value that represent a state}

\item{year}{an character or integer value that represent a year}
}
\value{
a plot of a map where dots represent accidents
}
\description{
Create a map of the count of accidents by giving information of year and state number
}
\details{
if state number does not exist, a warning message appears. In addition, function does return nothing if there is not information of accidents
}
