% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_functions_files.R
\name{fars_read}
\alias{fars_read}
\title{Read csv files}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{name (character) of the csv file}
}
\value{
an object of tbl_df class
}
\description{
Read csv files
}
\details{
if file does not exist, a warning message appears
}
