% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_functions_files.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read information of specific year}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{an character or integer value that represent a year}
}
\value{
an object of tbl_df class with two variables: month and year
}
\description{
Read information of specific year
}
\details{
if the value of year is not in the data, a warning message appears
}
