% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_functions_files.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Read information of specific year giving summarize information}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{an character or integer value that represent a year}
}
\value{
an object of tbl_df class with two variables: year and n (count)
}
\description{
Read information of specific year giving summarize information
}
\examples{
fars_summarize_years("2015")
fars_summarize_years(2015)
}
