% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_functions_files.R
\name{make_filename}
\alias{make_filename}
\title{Create a file name containing a specific year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{an character or integer value that represent a year}
}
\value{
an object of class character
}
\description{
Create a file name containing a specific year
}
\examples{
make_filename("2015")
make_filename(2015)
}
